refactor(memory): move private world runtime modules
