+++ conversation_memory_admin.py
@@ -25,7 +25,7 @@ from conversation_memory_port import (
     ConversationMemoryRecord,
     ConversationMemoryStatus,
 )
-from conversation_memory_identity import (
+from runtime.memory.conversation_memory_identity import (
     ConversationMemoryIdentityError,
     normalize_conversation_memory_user_id,
 )
