+++ conversation_memory_delivery.py
@@ -19,7 +19,7 @@ from conversation_memory_port import (
     MemoryWriteResult,
     MemoryWriteStatus,
 )
-from conversation_memory_identity import (
+from runtime.memory.conversation_memory_identity import (
     ConversationMemoryIdentityError,
     normalize_conversation_memory_user_id,
 )
