+++ mem0_memory.py
@@ -31,7 +31,7 @@ from conversation_memory_port import (
     NullConversationMemoryPort,
     UnavailableConversationMemoryPort,
 )
-from conversation_memory_identity import (
+from runtime.memory.conversation_memory_identity import (
     ConversationMemoryIdentityError,
     normalize_conversation_memory_user_id,
 )
